pyproject.toml
[project]
name = "sat-sentinel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "dask>=2023.10.0",
  "distributed>=2023.10.0",
  "geopandas>=0.14.0",
  "ipython>=9.10.0",
  "pandas<2.2.0",
  "pyarrow>=14.0.0",
  "pystac-client>=0.8.0",
  "pyyaml>=6.0.1",
  "rasterio<1.4.0",
  "rioxarray>=0.15.0",
  "shapely>=2.0.0",
  "xarray>=2023.10.0",
  "numpy<2.0.0",
  "bokeh>=3.8.2",
  "plotly>=6.5.2",
  "zarr>=3.1.5",
  "matplotlib>=3.10.8",
  "osmnx>=2.1.0",
  "folium>=0.20.0",
  "mapclassify>=2.10.0",
  "streamlit>=1.55.0",
  "streamlit-folium>=0.26.2",
  "odc-stac>=0.5.2",
  "scikit-image>=0.26.0",
  "segment-geospatial>=1.1.0",
  "duckdb>=1.5.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
